+++ example/a.ts
@@ -1,18 +1,12 @@
 import { Elysia, t } from '../src'
 
-new Elysia()
-	.macro({
-		token: {
-			resolve: () => {
-				return {
-					__token: '123'
-				}
-			}
-		}
-	})
-	.macro('some', {
-		token: true,
-		beforeHandle: ({ __token }) => {
-			console.log('__token', __token)
-		}
-	})
+new Elysia({ prefix: "/a" })
+	.get(
+    "/todo/:id",
+    ({ params }) => {
+        // O `params.id` aqui já é totalmente tipado!
+      return {
+        id: params.id,
+      };
+    }
+  )